Split out the House component and added color scss variables
