+++ src/Components/Main.scss
@@ -13,7 +13,7 @@
 @include foundation-card;
 @include foundation-button;
 
-
+//Variables for colours to be updated when there is a palette
 $color-red: darkred;
 $color-black: black;
 $color-white: white;
